+++ src/Admin/Views/AutofillTriage/Details.cshtml
@@ -1,7 +1,7 @@
 @using System.Net
 @using System.Text.Json
 @using Bit.Admin.Models
-@model Bit.Core.Entities.AutofillTriageReport
+@model Bit.Core.Autofill.Entities.AutofillTriageReport
 @{
     ViewData["Title"] = "Autofill Triage Report";
     var fields = new List<AutofillTriageFieldResultModel>();
