+++ src/Infrastructure.EntityFramework/Autofill/Models/AutofillTriageReport.cs
@@ -1,0 +1,18 @@
+﻿#nullable enable
+
+using AutoMapper;
+
+namespace Bit.Infrastructure.EntityFramework.Autofill.Models;
+
+// ReSharper disable once ClassWithVirtualMembersNeverInherited.Global
+public class AutofillTriageReport : Core.Autofill.Entities.AutofillTriageReport
+{
+}
+
+public class AutofillTriageReportMapperProfile : Profile
+{
+    public AutofillTriageReportMapperProfile()
+    {
+        CreateMap<Core.Autofill.Entities.AutofillTriageReport, AutofillTriageReport>().ReverseMap();
+    }
+}
